Fix: minor issues in Docker Compose manifest (#51)
As the building of the notebook Docker image can take a while, we have added
another manifest to do with the published version.

Also the host domain `host.docker.internal` is unknown. One option is to list
it as an extra host. Another option, the one applied here, is to add all the
services to the same Docker network used by the backend services.

Finally, the variable `HEXA_SERVER_URL` is missing and should be passed to the
JupyterHub cotnainer.

Co-authored-by: Christophe Philemotte <1059797+toch@users.noreply.github.com>