Fix: failing notebook start when local (#50)
When running locally the JupyterHub, we can't spawn a notebook. In the logs,
the container creation returns a 204 HTTP status, i.e. created, but the
container can't be found when listing them.

After having checked that the host Docker engine is accessible, it appears that
the container was crashing immediatly. We could confirm that by reproducing the
running command use to start the container. It was retrieved thanks to
@pvanliefland. This command is `/bin/bash -o pipefail -c singleuser`. But the
entrypoint of the notebook expects the one of the following command: `notebook`,
`singleuser`, `pipeline`, or `help`.

According to the [JupyterHub DockerSpawner document](https://jupyterhub-dockerspawner.readthedocs.io/en/latest/api/index.html#dockerspawner.DockerSpawner.image),
it's possible to change the command passed to run a container by setting
`c.Spawner.cmd` in the config. That's the fix.

Co-authored-by: Christophe Philemotte <1059797+toch@users.noreply.github.com>